update - various fixes, naming, ordering, etc.
