+++ frontend/pages/Home.js
@@ -1,11 +1,7 @@
 import React from "react";
-import Sections from "../components/Sections";
-import "../scss/animation.scss";
-
-// animation
+import HomeSections from "../components/HomeSections";
 import Logo_Gif from "../images/gif/rec-animation-english---Copy.gif";
 import Car from "../images/featured/BHVP_PH_CetaitParis_DP_0122_01_008.jpg";
-
 
 
 export class HomePage extends React.Component {
@@ -14,17 +10,14 @@ export class HomePage extends React.Component {
     }
 
     render() {
-
         return (
-            <div>
-                <div className="animation" style={{backgroundImage: `url(${Logo_Gif})`}}>
-                    <img src={Car} className="background"/>
-
+            <section>
+                <div className="main-section">
+                    <img src={Logo_Gif} alt="Paris Logo"/>
+                    <img src={Car} className="background" alt="Background image: Car"/>
                 </div>
-
-
-                <Sections/>
-            </div>
+                <HomeSections/>
+            </section>
         );
     }
 }
