+++ frontend/scss/home-sections.scss
@@ -1,0 +1,104 @@
+#home-sections {
+    a {
+        text-decoration: none;
+        color: black;
+    }
+
+    a:focus {
+        color: white;
+    }
+
+    a:active {
+        color: #5E5F5F;
+    }
+
+    .large-arrow {
+        position: relative;
+        font-size: 72px;
+        font-family: "Segoe UI Symbol";
+        font-weight: 100;
+        margin-left: 349px;
+
+    }
+
+    .photo-archive {
+        background-color: #00BFE1;
+        max-height: 397px;
+        column-rule: black 1px solid;
+        box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
+
+    }
+
+    .map-section {
+        background-color: #FF6C6C;
+        max-height: 397px;
+        column-rule: black 1px solid;
+        box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+
+    }
+
+    .context {
+        background-color: rgba(95, 90, 98, 0.29);
+        max-height: 397px;
+        column-rule: black 1px solid;
+        box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
+
+    }
+
+
+    .text {
+        text-align-all: center;
+        border-inline: 1px solid black;
+        padding-top: 45px;
+        padding-left: 37px;
+
+        h5 {
+            margin-bottom: 15px;
+            font-family: $font-family-bold;
+        }
+
+        p {
+            font-weight: normal;
+            font-size: 24px;
+            margin-bottom: 150px;
+            font-family: $font-family-base;
+        }
+
+        i {
+            position: relative;
+            font-size: 20px;
+
+        }
+    }
+
+    .bi {
+        margin-left: 170px;
+    }
+
+
+    .photo {
+        //height: 397px;
+        background-position: center;
+        background-repeat: no-repeat;
+        background-size: cover;
+        overflow: hidden;
+        mix-blend-mode: multiply;
+    }
+
+    @media screen and (max-width: 767px) {
+        .text {
+            p {
+                font-size: 20px;
+                margin-bottom: 30px;
+            }
+        }
+    }
+
+    @media screen and (min-width: 767px) {
+        .text {
+            i {
+                display: none;
+            }
+        }
+    }
+}
