+++ frontend/scss/home.scss
@@ -122,8 +122,3 @@ code {
     stroke: #FF3D3A;
     fill: #FF3D3A;
 }
-
-.home-map {
-    width: 100%;
-    height: 500px !important;
-}
