+++ frontend/scss/styles.scss
@@ -18,6 +18,8 @@ $footer-height: 150px;
 @import './fonts';
 @import './variables';
 @import './home';
+@import './home-sections';
+@import './home-main-section';
 @import './about';
 @import './photoView';
 @import './photographerView';
@@ -30,8 +32,6 @@ $footer-height: 150px;
 @import './legend';
 @import './parisMap';
 @import './mapPageView';
-@import './sections';
-@import './animation';
 @import './tags';
 @import './photographerListView';
